+++ tests/unit/workflows/pipeline_store/test_install.py
@@ -68,7 +68,7 @@ def workflow(
     )
 
     # mock singularity/apptainer pull (this is overridden by some tests)
-    mocker.patch("nipoppy.workflows.pipeline_store.install._run_command")
+    mocker.patch("nipoppy.workflows.pipeline_store.install.run_command")
 
     return workflow
 
@@ -241,7 +241,7 @@ def test_download_container(
         return_value=ApptainerHandler(),
     )
     mocked_run_command = mocker.patch(
-        "nipoppy.workflows.pipeline_store.install._run_command"
+        "nipoppy.workflows.pipeline_store.install.run_command"
     )
 
     workflow._download_container(pipeline_config)
@@ -285,7 +285,7 @@ def test_download_container_confirm_true(
     )
 
     mocked_run_command = mocker.patch(
-        "nipoppy.workflows.pipeline_store.install._run_command"
+        "nipoppy.workflows.pipeline_store.install.run_command"
     )
 
     workflow._download_container(pipeline_config)
@@ -318,7 +318,7 @@ def test_download_container_status(
         f"nipoppy.workflows.pipeline_store.install.{console}.status",
     )
     mocked_run_command = mocker.patch(
-        "nipoppy.workflows.pipeline_store.install._run_command"
+        "nipoppy.workflows.pipeline_store.install.run_command"
     )
 
     workflow.study.config.CONTAINER_CONFIG.COMMAND = command
@@ -339,7 +339,7 @@ def test_download_container_failed(
 ):
     error_message = "Download failed"
     mocked = mocker.patch(
-        "nipoppy.workflows.pipeline_store.install._run_command",
+        "nipoppy.workflows.pipeline_store.install.run_command",
         side_effect=subprocess.CalledProcessError(1, error_message),
     )
 
@@ -360,7 +360,7 @@ def test_download_container_no_uri(
     mocker: pytest_mock.MockFixture,
 ):
     pipeline_config.CONTAINER_INFO.URI = None
-    mocked = mocker.patch("nipoppy.workflows.pipeline_store.install._run_command")
+    mocked = mocker.patch("nipoppy.workflows.pipeline_store.install.run_command")
 
     workflow._download_container(pipeline_config)
 
@@ -378,7 +378,7 @@ def test_download_container_image_exists(
     )
     fpath_container.parent.mkdir(parents=True, exist_ok=True)
     fpath_container.touch()
-    mocked = mocker.patch("nipoppy.workflows.pipeline_store.install._run_command")
+    mocked = mocker.patch("nipoppy.workflows.pipeline_store.install.run_command")
 
     workflow._download_container(pipeline_config)
     mocked.assert_not_called()
